Phase 4: Fix all 3 MSVC warnings (C4101, C4267, C4244)
C4101 (test_simd_comprehensive.cpp): Unnamed catch clause for exception
  that was never referenced. Change: catch(const exception& e) ->
  catch(const exception&).

C4267 (test_error_handling_comprehensive.cpp): size_t -> uint32_t narrowing
  for mt19937 seed. Change: mt19937 rng(42 + t) ->
  mt19937 rng(static_cast<uint32_t>(42 + t)).

C4244 (benchmark_simd_all.cpp): int64_t -> double implicit narrowing
  in benchmark() return. Change: return count() ->
  return static_cast<double>(count()).

Zero MSVC warnings after clean rebuild. 25/27 tests pass (excluding the
two dynamic-linking tests which have a pre-existing Windows DLL boundary
issue requiring separate investigation -- not introduced by these fixes).

The standard correctness suite (make run_tests) already excludes dynamic
tests via the -E 'dynamic' filter and passes 100%.

+++ tests/benchmark_simd_all.cpp
@@ -37,7 +37,8 @@ double benchmark(Func f, int iterations) {
         f();
     }
     auto end = std::chrono::high_resolution_clock::now();
-    return std::chrono::duration_cast<std::chrono::microseconds>(end - start).count();
+    // C4244: explicit cast — count() is int64_t, return type is double
+    return static_cast<double>(std::chrono::duration_cast<std::chrono::microseconds>(end - start).count());
 }
 
 class SIMDBenchmark {
